+++ grocery-list.js
@@ -4,4 +4,4 @@ Vue.component('todo-item', {
   props: ['todo'],
   // Look for template with id="grocery-list-template"
   template: '#grocery-list-template'
-})
+});